refactor htmls based on changes in domain project

+++ src/main/java/ir/niopdc/station/controller/mvc/ZoneViewController.java
@@ -5,6 +5,7 @@ import ir.niopdc.domain.zone.Zone;
 import ir.niopdc.domain.zone.ZoneService;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.MessageSource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -12,6 +13,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Locale;
 
 @Controller
 @RequiredArgsConstructor
@@ -19,6 +21,7 @@ import java.util.List;
 public class ZoneViewController {
 
     private final ZoneService zoneService;
+    private final MessageSource messageSource;
 
     @GetMapping("/get-all-zones")
     public String getAllZones(Model model) {
@@ -34,12 +37,12 @@ public class ZoneViewController {
     }
 
     @PostMapping("/create")
-    public ResponseEntity<String> createZone(@ModelAttribute Zone zone) {
+    public ResponseEntity<String> createZone(@ModelAttribute Zone zone, Locale locale) {
         try {
             zoneService.save(zone);
-            return ResponseEntity.ok("منطقه با موفقیت اضافه شد!");
+            return ResponseEntity.ok(messageSource.getMessage("zone_created_successfully", null, locale));
         } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("خطا در اضافه کردن منطقه!");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(messageSource.getMessage("error_creating_zone", null, locale));
         }
     }
 
@@ -63,33 +66,37 @@ public class ZoneViewController {
     }
 
     @PostMapping("/edit")
-    public ResponseEntity<Model> updateZone(@ModelAttribute Zone zone, Model model) {
+    public ResponseEntity<Model> updateZone(@ModelAttribute Zone zone, Model model, Locale locale) {
         try {
             Zone existingZone = zoneService.findById(zone.getId());
             if (existingZone != null) {
                 existingZone.setName(zone.getName());
                 Zone updatedZone = zoneService.update(existingZone);
                 model.addAttribute("zone", updatedZone);
-                model.addAttribute("message", "منطقه با موفقیت ویرایش شد!");
+                model.addAttribute("message", messageSource.getMessage("zone_edited_successfully", null, locale));
                 return ResponseEntity.ok(model);
             } else {
-                model.addAttribute("message", "منطقه پیدا نشد!");
+                model.addAttribute("message", messageSource.getMessage("zone_not_found", null, locale));
                 return ResponseEntity.status(HttpStatus.NOT_FOUND).body(model);
             }
         } catch (Exception e) {
-            model.addAttribute("message", "خطا در ویرایش منطقه!");
+            model.addAttribute("message", messageSource.getMessage("error_editing_zone", null, locale));
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(model);
         }
     }
 
-
     @DeleteMapping("/delete")
-    public ResponseEntity<String> deleteZone(@RequestParam("zoneId") Long zoneId) {
+    public ResponseEntity<String> deleteZone(@RequestParam("zoneId") Long zoneId, Locale locale) {
         try {
+            Zone zone = zoneService.findById(zoneId);
+            if (zone.getAreas() != null && !zone.getAreas().isEmpty()) {
+                return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                        .body(messageSource.getMessage("zone_has_related_areas", null, locale)); // Custom message for zones with areas
+            }
             zoneService.deleteById(zoneId);
-            return ResponseEntity.ok("منطقه با موفقیت حذف شد!");
+            return ResponseEntity.ok(messageSource.getMessage("zone_deleted_successfully", null, locale));
         } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("خطا در حذف منطقه!");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(messageSource.getMessage("error-deleting_zone", null, locale));
         }
     }
 }
